fix parsing bug related to feeds
Reported by (Kevin Colyer)++.

Feed operators can not go on the next line if the line they follow end
in a closing curly brace. This is explained in S04's section
"Statement-ending blocks" -- basically, if a closing curly at the end
of a line *can* end a statement, it does.